top_shows/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cssselect==1.1.0
dataclasses==0.8
Django==3.2.7
djangorestframework==3.12.4
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
PyMySQL==1.0.2
pyquery==1.4.3
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.6.0
